雷とステージの足場の反射を作り途中

+++ Project1/HitBox.h
@@ -10,7 +10,7 @@ class HitBox
 private:
 	int px0, px1, py0, py1;
 	int ex0, ex1, ey0, ey1;
-	int tx0, tx1, ty0, ty1;
+	float tx0, tx1, ty0, ty1;
 	int fx0, fx1, fy0, fy1;
 	
 	int underPx0, underPy0, underPx1, underPy1;	//プレイヤーの接地判定
